User specific Database modified

+++ todo_app/views.py
@@ -42,23 +42,27 @@ def index(request):
             fm=Add_Task()
     elif request.method=='GET':
         fm=Add_Task()'''
-    T=Tasks.objects.all()
+    T=Tasks.objects.filter(userid=request.user.id)
     #calculate total tasks done today from Tasks and Recurring
     #delete todays task data from Progress
     #add new data with todays date and progress value stored in p
-    p=len(Recurring.objects.filter(done=1)) + len(Tasks.objects.filter(done=1))
-    ttl=len(Recurring.objects.all()) + len(Tasks.objects.all())
+    p=len(Recurring.objects.filter(userid=request.user.id,done=1)) + len(Tasks.objects.filter(userid=request.user.id,done=1))
+    ttl=len(Recurring.objects.filter(userid=request.user.id)) + len(Tasks.objects.filter(userid=request.user.id))
+    
     #in try: check if todays data is available or not, if yes then delete the data.
     try:
-     d = Progress.objects.filter(date=tme)
+     d = Progress.objects.filter(userid=request.user.id,date=tme)
      d.delete()
     except:
         pass
     #enter new data of today
     #update progress
-    obj = Progress(date=tme,progress=p,total=ttl)
-    obj.save()
-     
+    try:
+        obj = Progress(date=tme,progress=p,total=ttl,userid=request.user)
+        obj.save()
+    except:
+        pass
+ 
 
 
     return render(request,'index.html',{'time':date,'T':T,'H':h})
@@ -75,7 +79,8 @@ def add(request):
         if fm.is_valid():
             fm.save()
             fm=Add_Task() 
-    T=Tasks.objects.all()
+
+    T=Tasks.objects.filter(userid=request.user.id)
     return render(request,'index.html',{'time':date,'T':T,'form':fm,'H':h})
 
 def delete(request,id):
@@ -97,7 +102,7 @@ def update(request,id):
         h=0
         u = Tasks.objects.get(pk=id)
         upd= Add_Task(instance=u)
-    T=Tasks.objects.all()
+    T=Tasks.objects.filter(userid=request.user.id)
     tiu=Tasks.objects.get(pk=id)
     return render(request,'update.html',{'time':date,'T':T,'form':upd,'H':h,'ti':tiu})
 
@@ -115,13 +120,13 @@ def finished(request,id):
 
 def delete_all(request):
     if request.method=='GET':
-        T=Tasks.objects.all()
+        T=Tasks.objects.filter(userid=request.user.id)
         T.delete()
         return HttpResponseRedirect('/')
 
 def clear_finished(request):
     if request.method=='GET':
-        T=Tasks.objects.filter(done=1)
+        T=Tasks.objects.filter(userid=request.user.id,done=1)
         T.delete()
         return HttpResponseRedirect('/')
 
@@ -131,14 +136,17 @@ def clear_finished(request):
 
 def recurring(request):
     #upddate progress
-    p=len(Recurring.objects.filter(done=1)) + len(Tasks.objects.filter(done=1))
+    if not request.user.is_authenticated:
+            return redirect('account/login')
+    p=len(Recurring.objects.filter(userid=request.user.id,done=1)) + len(Tasks.objects.filter(userid=request.user.id,done=1))
+    ttl=len(Recurring.objects.filter(userid=request.user.id)) + len(Tasks.objects.filter(userid=request.user.id))
     try:
-     d = Progress.objects.filter(date=tme)
+     d = Progress.objects.filter(userid=request.user.id,date=tme)
      d.delete()
     except:
         pass
     #update progress
-    obj = Progress(date=tme,progress=p)
+    obj = Progress(date=tme,progress=p,total=ttl,userid=request.user)
     obj.save()
     
     #set done=0, if date stored in data's != today's date
@@ -151,7 +159,7 @@ def recurring(request):
     if len(T)==0:
         Recurring.objects.update(done = 0)
         Recurring.objects.update(date=todays_date)
-    T=Recurring.objects.all()
+    T=Recurring.objects.filter(userid=request.user.id)
     return render(request,'Recurring.html',{'time':date,'T':T,'H':h})
 
 def R_add(request):
@@ -164,7 +172,7 @@ def R_add(request):
         if fm.is_valid():
             fm.save()
             fm=R_Add_Task() 
-    T=Recurring.objects.all()
+    T=Recurring.objects.filter(userid=request.user.id)
     return render(request,'Recurring.html',{'time':date,'T':T,'form':fm,'H':h})
 
 def R_delete(request,id):
@@ -186,7 +194,7 @@ def R_update(request,id):
         h=0
         u = Recurring.objects.get(pk=id)
         upd= R_Add_Task(instance=u)
-    T=Recurring.objects.all()
+    T=Recurring.objects.filter(userid=request.user.id)
     tiu=Recurring.objects.get(pk=id)
     return render(request,'update.html',{'time':date,'T':T,'form':upd,'H':h,'ti':tiu})
 
@@ -204,20 +212,29 @@ def R_finished(request,id):
 
 def R_delete_all(request):
     if request.method=='GET':
-        T=Recurring.objects.all()
+        T=Recurring.objects.filter(userid=request.user.id)
         T.delete()
         return HttpResponseRedirect('/recurring')
 
 def R_clear_finished(request):
     if request.method=='GET':
-        T=Recurring.objects.filter(done=1)
+        T=Recurring.objects.filter(userid=request.user.id,done=1)
         T.delete()
         return HttpResponseRedirect('/recurring')
 
 def progress(request):
     if request.method=='GET':
-        T=Progress.objects.all().order_by('id')
-        C=Chartcolor.objects.all()
+        #if user is not logged in, then take user to login page.
+        if not request.user.is_authenticated:
+            return redirect('account/login')
+        
+        T=Progress.objects.filter(userid=request.user.id).order_by('id')
+        C=Chartcolor.objects.filter(userid=request.user.id)
+        if len(C)==0:
+            Chartcolor.objects.create(color='line',userid=request.user)
+        C=Chartcolor.objects.filter(userid=request.user.id)
+
+
         xValues=[]
         yValues=[]
         zValues=[]
@@ -229,44 +246,53 @@ def progress(request):
         xValues=xValues[-7:]
         yValues=yValues[-7:]
         zValues=zValues[-7:]
+       
         return render(request,'progress.html',{'time':date,'xVal':xValues,'yVal':yValues,'zVal':zValues,'color':C[0].color}) 
 
 def changechartcolor(request,color):
     if request.method=='GET':
-        C=Chartcolor.objects.all()
-        C.delete()
-        upd=Chartcolor(color=color)
-        upd.save()
+        Chartcolor.objects.filter(userid=request.user.id).update(color=color)
         return HttpResponseRedirect('/progress')
-
 
 def changeaudio(request,audio_number):
     if request.method=='GET':
-        C=Audio.objects.all()
-        C.delete()
-        upd=Audio(audio_number=audio_number)
-        upd.save()
+        Audio.objects.filter(userid=request.user.id).update(audio_number=audio_number)
         return HttpResponseRedirect('/timer')
 
 def timer(request):
     if request.method=='GET':
+        #if user is not logged in, take them to login page.
+        if not request.user.is_authenticated:
+            return redirect('account/login')
         #fetch the value of set timer from Timer table
-        T=Timer.objects.all()
+        T=Timer.objects.filter(userid=request.user.id)
+
+        #if logged in user data is not present in Timer and Audio table, then insert it
+        if len(T)==0:
+            Timer.objects.create(time=120,userid=request.user)
+            Audio.objects.create(audio_number=1,userid=request.user)
+
+        T=Timer.objects.filter(userid=request.user.id)
+
         #lets store all audio in a file:
         #any from database we will get the index and that index audio will be sent to timer page.
         aud_list = ['TGSZM2N-old-alarm-clock.mp3','y2mate-com-scam-1992-bgm-52384.mp3','y2mate-com-harry-potter-ringtone-bgm-tone-54095.mp3','sultan-abdul-hamid-music125165-52996.mp3','kgf-bgm-ringtone-44262.mp3','titanic-20romantic-23691.mp3']
-        #trying to access the contents of Audio_store
+        #trying to access the media files uploaded by user from media folder
         #media_path = settings.MEDIA_ROOT
         #myfiles = [f for f in listdir(media_path) if isfile(join(media_path, f))]
+        #trim the name of audio files in myfiles
         global custom_audio
-        myfiles = Audio_store.objects.all()
+        myfiles = Audio_store.objects.filter(userid=request.user.id)
         #trim the name of audio files in myfiles
         custom_audio = myfiles
+
         custom_aud=[]
         for i in myfiles:
             custom_aud.append('media/' + str(i.record))
             
-        index=Audio.objects.all()[0].audio_number
+
+        index=Audio.objects.filter(userid=request.user.id)[0].audio_number
+        
         try:
           aud = staticfiles_storage.url(aud_list[index])
         except:
@@ -275,13 +301,10 @@ def timer(request):
         temp=[]
         global vmp
         vmp=myfiles
-        
         for i in custom_aud:
             a=i[:-4]
             temp.append(a[16:30] if len(a)>12 else a)
-
         myfiles=temp
-        #myfiles=custom_aud
         audio_name =['Bell','Scam 1992', 'Harry Potter','Drive Forever','KGF','Titanic']
         audio_name+=myfiles
         #Time is in seconds
@@ -313,11 +336,11 @@ def timer(request):
         #calculate in seconds
         tme=hr*60*60 + min*60 + sec
 
-        print(tme)
+        
         #push this data in Timer table
-        T=Timer.objects.all()
+        T=Timer.objects.filter(userid=request.user.id)
         T.delete()
-        T=Timer(time=tme)
+        T=Timer(time=tme,userid=request.user)
         T.save()
         return HttpResponseRedirect('/timer')
       
@@ -331,15 +354,12 @@ def simple_upload(request):
         #provide limit of 3 files to be added, if more than 3 is addded, pass a message to user
         #media_path = settings.MEDIA_ROOT
         #myfiles = [f for f in listdir(media_path) if isfile(join(media_path, f))]
-        #retrieve data from Audio_stores
         myfiles = Audio_store.objects.all()
-        #print(myfiles[0].record)
         if myfile.name[-3:] == 'mp3':
             if len(myfiles) <=2 :
                 #fs = FileSystemStorage()
                 #filename = fs.save(myfile.name, myfile)
-                #save the myfile in Audio_store
-                upd=Audio_store(record=myfile)
+                upd=Audio_store(record=myfile,userid=request.user)
                 upd.save()
             else:
                 messages.info(request,'There is a limit of 3 custom audio files. Delete 1 to add this.')
@@ -354,6 +374,6 @@ def delete_audio(request,num):
     changeaudio(request,0)
     #now delete the audio number passed from media folder
     #default_storage.delete('/Users/DCQUASTER JACK/projects/todo/media/'+vmp[6-num])
-    d = Audio_store.objects.get(record=custom_audio[num-6].record)
+    d = Audio_store.objects.get(record=custom_audio[num-6].record,userid=request.user.id)
     d.delete()
     return HttpResponseRedirect('/timer')
